First Commit This add 2 bash script
THIS CODE REQUIRE CURL
(apt-get install curl)

--> create_channel username
Create a youtube channel local repository

It will first create a directory username
and add an index.html files inside it
It will then download the last video of the user and create a last.txt file
with the link of the last video.

It will also create a cron who will seek the last video every 5 minutes
(WARNING if you activate email on the server, it will send you an email every 5 minutes
which will undoubtly ban your ip from your smtp relay!

--> get_last username
Get_last video of username channel
It will first check if you have already download the last video on the channel
If you haven't it will download it and create a link on the index.html of the channel to access it
IT WILL NOT WORK IF THERE ISN'T ANY HD VERSION (because it will not be MP4)
OBVIOUSLY IT WILL NOT WORK ON FIREFOX (since it an mp4 not a ogv theora)

Next version will add the possibility to convert it to ogv to have firefox/chrome compatibility

index.html is an page with all the channels

This set of script will work even if you're not using a webserver.